KAFKA-9804:Extract consumer configs out of PerfConfig
Configs for producer has been extracted out of PerfConfig in
https://github.com/apache/kafka/pull/3613/commits. The remaining
in PerfConfig are configs for consumer. And ConsumerPerformance
also has configs for consumer. Separating these configs into two
classes is not concise. So we can put all configs into class
ConsumerPerformance.ConsumerPerfConfig.
